backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv
langchain
openai
chromadb
pdfplumber
unstructured
pypandoc
pypdf
pandas
python-docx
openpyxl
docx2txt
aiofiles
httpx
gunicorn
sqlalchemy
pymysql
langchain-community
python-multipart
tiktoken
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]